# docs/swagger.yaml
basePath: /api/v1
info:
  contact:
    email: romber2001@gmail.com
    name: Romber Li
    url: http://www.github.com/romberli/das
  description: das is shortcut for Database Autonomy Service, it provides self-perception,
    repair, optimization, and security database service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DAS
  version: "1.0"
paths:
...
  /api/v1/metadata/middleware-server:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"last_update_time":"2021-11-18T15:54:10.599097+08:00","cluster_id":1,"server_name":"middleware-server-1","port_num":33061,"del_flag":0,"create_time":"2021-11-17T14:47:10.521279+08:00","id":1,"middleware_role":1,"host_ip":"192.168.10.219"}]}
          schema:
            type: string
      summary: get all middleware servers
      tags:
      - middleware server
    post:
      consumes:
      - application/json
      parameters:
      - description: middleware cluster id
        in: body
        name: cluster_id
        required: true
        schema:
          type: integer
      - description: middleware server name
        in: body
        name: server_name
        required: true
        schema:
          type: string
      - description: middleware role
        in: body
        name: middleware_role
        required: true
        schema:
          type: integer
      - description: host ip
        in: body
        name: host_ip
        required: true
        schema:
          type: string
      - description: port number
        in: body
        name: port_num
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"host_ip":"192.168.10.219","del_flag":0,"create_time":"2022-03-02T10:18:28.021994+08:00","last_update_time":"2022-03-02T10:18:28.021994+08:00","cluster_id":1,"server_name":"new_middleware_server","middleware_role":1,"id":42,"port_num":33062}]}
          schema:
            type: string
      summary: add a new middleware server
      tags:
      - middleware server
  /api/v1/metadata/middleware-server/cluster-id/:cluster_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: middleware cluster id
        in: path
        name: cluster_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"id":1,"cluster_id":1,"server_name":"middleware-server-1","host_ip":"192.168.10.219","port_num":33061,"del_flag":0,"create_time":"2021-11-17T14:47:10.521279+08:00","last_update_time":"2021-11-18T15:54:10.599097+08:00","middleware_role":1}]}
          schema:
            type: string
      summary: get middleware servers by cluster id
      tags:
      - middleware server
  /api/v1/metadata/middleware-server/delete/:id:
    post:
      consumes:
      - application/json
      parameters:
      - description: middleware server id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"server_name":"new_middleware_server","host_ip":"192.168.10.219","port_num":33062,"create_time":"2022-03-02T10:18:28.021994+08:00","last_update_time":"2022-03-02T10:18:28.021994+08:00","id":42,"cluster_id":1,"middleware_role":1,"del_flag":0}]}
          schema:
            type: string
      summary: delete middleware server by id
      tags:
      - middleware server
  /api/v1/metadata/middleware-server/get/:id:
    get:
      consumes:
      - application/json
      parameters:
      - description: middleware server id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"last_update_time":"2021-11-18T15:54:10.599097+08:00","id":1,"server_name":"middleware-server-1","middleware_role":1,"port_num":33061,"cluster_id":1,"host_ip":"192.168.10.219","del_flag":0,"create_time":"2021-11-17T14:47:10.521279+08:00"}]}
          schema:
            type: string
      summary: get middleware server by id
      tags:
      - middleware server
  /api/v1/metadata/middleware-server/host-info:
    get:
      consumes:
      - application/json
      parameters:
      - description: host ip
        in: body
        name: host_ip
        required: true
        schema:
          type: string
      - description: port number
        in: body
        name: port_num
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"middleware_role":1,"host_ip":"192.168.10.219","del_flag":0,"last_update_time":"2021-11-18T15:54:10.599097+08:00","id":1,"server_name":"middleware-server-1","cluster_id":1,"port_num":33061,"create_time":"2021-11-17T14:47:10.521279+08:00"}]}
          schema:
            type: string
      summary: get middleware server by host info
      tags:
      - middleware server
  /api/v1/metadata/middleware-server/update/:id:
    post:
      consumes:
      - application/json
      parameters:
      - description: middleware cluster id
        in: body
        name: cluster_id
        required: true
        schema:
          type: integer
      - description: middleware server name
        in: body
        name: server_name
        schema:
          type: string
      - description: middleware role
        in: body
        name: middleware_role
        schema:
          type: integer
      - description: host ip
        in: body
        name: host_ip
        schema:
          type: string
      - description: port number
        in: body
        name: port_num
        schema:
          type: integer
      - description: delete flag
        in: body
        name: del_flag
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: middleware_servers":[{"del_flag":0,"server_name":"update_middeware_server","host_ip":"192.168.10.219","port_num":33061,"middleware_role":1,"create_time":"2021-11-17T14:47:10.521279+08:00","last_update_time":"2021-11-18T15:54:10.599097+08:00","id":1,"cluster_id":1}]}
          schema:
            type: string
      summary: update middleware server by id
      tags:
      - middleware server
...
swagger: "2.0"